+++ internal/common/autogen/import.go
@@ -1,4 +1,4 @@
-package autogeneration
+package autogen
 
 import (
 	"context"
