+++ internal/common/autogen/import_test.go
@@ -1,10 +1,10 @@
-package autogeneration_test
+package autogen_test
 
 import (
 	"fmt"
 	"testing"
 
-	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/autogeneration"
+	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/autogen"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
@@ -40,13 +40,13 @@ func TestGenericImportOperation(t *testing.T) {
 			name:          "Error: Wrong number of attributes",
 			importID:      "5c9d0a239ccf643e6a35ddasdf/myCluster",
 			idAttributes:  []string{"project_id", "cluster_name", "region"},
-			expectedError: conversion.StringPtr(fmt.Sprintf(autogeneration.ExpectedErrorMsg, "project_id/cluster_name/region")),
+			expectedError: conversion.StringPtr(fmt.Sprintf(autogen.ExpectedErrorMsg, "project_id/cluster_name/region")),
 		},
 	}
 
 	for _, tc := range testCases {
 		t.Run(tc.name, func(t *testing.T) {
-			attrValues, err := autogeneration.ProcessImportID(tc.importID, tc.idAttributes)
+			attrValues, err := autogen.ProcessImportID(tc.importID, tc.idAttributes)
 
 			if tc.expectedError != nil {
 				require.Error(t, err)
