+++ internal/common/autogen/unmarshal.go
@@ -1,4 +1,4 @@
-package autogeneration
+package autogen
 
 import (
 	"context"
